Some changes to speech retrieval to merge into the main branch
