--- examples/flickr40k_retrieval.yaml
@@ -1,51 +1,0 @@
-defaults:
-    experiment:
-        model_file: examples/output/<EXP>.mod
-        hyp_file: examples/output/<EXP>.hyp
-        out_file: examples/output/<EXP>.out
-        err_file: examples/output/<EXP>.err
-        run_for_epochs: 20
-        eval_metrics: recall|nbest=1,recall|nbest=5,mean_avg_precision|nbest=5
-    train:
-        batch_strategy: 'shuffle'
-        trainer: 'msgd'
-        learning_rate: 0.00001
-        momentum: 0.9 
-        default_layer_dim: 512
-        batch_size: 128
-        reload_between_epochs: True    ### Comment this out if you can allocate enough memory to use the whole training corpus every epoch (and sample_train_sents below). This is necessary in interactive sessions because the memory is limited to 4GB, but in batch sessions this may not be necessary.
-        dropout: 0.0
-        dev_metrics: recall|nbest=1,recall|nbest=5
-        training_corpus: !BilingualTrainingCorpus
-            train_src: /pylon5/ci560op/gneubig/data/flickr/flickr40k_speech_train.npz
-            train_trg: /pylon5/ci560op/gneubig/data/flickr/flickr40k_ids_train.txt
-            dev_src: /pylon5/ci560op/gneubig/data/flickr/flickr40k_speech_val.npz
-            dev_trg: /pylon5/ci560op/gneubig/data/flickr/flickr40k_ids_val.txt
-        corpus_parser: !BilingualCorpusParser
-            src_reader: !ContVecReader
-                transpose: True
-            trg_reader: !IDReader {}
-            sample_train_sents: 5000  ### Similarly, comment this out if reload_between_epochs is commented out
-        model: !DotProductRetriever
-            src_embedder: !NoopEmbedder
-                emb_dim: 40
-            src_encoder: !HarwathSpeechEncoder
-                filter_height: [40, 1, 1]
-                filter_width: [5, 25, 25]
-                channels: [1, 64, 512]
-                num_filters: [64, 512, 1024]
-                stride: [1, 1, 1]
-            trg_embedder: !NoopEmbedder
-                emb_dim: 40
-            trg_encoder: !HarwathImageEncoder
-                in_height: 4096
-                out_height: 1024 
-            database: !StandardRetrievalDatabase
-                reader: !ContVecReader {}
-                database_file: /pylon5/ci560op/gneubig/data/flickr/flickr40k_images_all.npz
-    decode:
-        src_file: /pylon5/ci560op/gneubig/data/flickr/flickr40k_speech_test.npz
-    evaluate:
-        ref_file: /pylon5/ci560op/gneubig/data/flickr/flickr40k_ids_test.npz
-
-speech_retriever_baseline:
