+++ xnmt/input.py
@@ -138,95 +138,25 @@ class ContVecReader(InputReader, Serializable):
   def __init__(self, transpose=True):
     self.transpose = transpose
   def read_sents(self, filename, filter_ids=None, in_npy_format=False):
-    if not in_npy_format:
-      npzFile = np.load(filename, mmap_mode=None if filter_ids is None else "r")
-      npzKeys = sorted(npzFile.files, key=lambda x: int(x.split('_')[-1]))
-      if filter_ids is not None:
-        npzKeys = [npzKeys[i] for i in filter_ids]
-      for idx, key in enumerate(npzKeys):
-        inp = npzFile[key]
-        if self.transpose:
-          inp = inp.transpose()
-        if idx % 1000 == 999:
-          print("Read {} lines ({:.2f}%) of {} at {}".format(idx+1, float(idx+1)/len(npzKeys)*100, filename, key))
-        yield ArrayInput(inp)
-      npzFile.close()
-    else:
-      npyFile = []
-      flag = 0
-      with open(filename, 'r') as f:
-        while not flag:
-          cur_file = f.readline().rstrip()
-          if not cur_file:
-            flag = 1
-            break
-          else: 
-            npyFile.append(cur_file)#np.load(cur_file)
-        if filter_ids is not None:
-          npyFile = [npyFile[i] for i in filter_ids]
-        for f in npyFile:
-          #print(f)
-          inp = np.load(f) 
-          if self.transpose:
-            inp = inp.transpose()
-          yield ArrayInput(inp)
-            
+    npzFile = np.load(filename, mmap_mode=None if filter_ids is None else "r")
+    npzKeys = sorted(npzFile.files, key=lambda x: int(x.split('_')[-1]))
+    if filter_ids is not None:
+      npzKeys = [npzKeys[i] for i in filter_ids]
+    for idx, key in enumerate(npzKeys):
+      inp = npzFile[key]
+      if self.transpose:
+        inp = inp.transpose()
+      if idx % 1000 == 999:
+        print("Read {} lines ({:.2f}%) of {} at {}".format(idx+1, float(idx+1)/len(npzKeys)*100, filename, key))
+      yield ArrayInput(inp)
+    npzFile.close()
 
   def count_sents(self, filename):
     npzFile = np.load(filename, mmap_mode="r") # for counting sentences, only read the index
     l = len(npzFile.files)
     npzFile.close()
-    '''l = 0
-    with open(filename) as f:
-      while True:
-        line = f.readline()
-        if not line:
-          break
-        else:
-          l += 1'''
     return l
 
-class ContVecNPYReader(InputReader, Serializable):
-  yaml_tag = u"!ContVecNPYReader"
-  def __init__(self, transpose=True):
-      self.transpose = transpose
-
-  def read_sents(self, filename, filter_ids=None):
-    npyFile = []
-    flag = 0
-    with open(filename, 'r') as f:
-      while not flag:
-        cur_file = f.readline().rstrip()
-        if not cur_file:
-          flag = 1
-          break
-        else:
-          npyFile.append(cur_file)#np.load(cur_file)
-      
-      if filter_ids is not None:
-        npyFile = [npyFile[i] for i in filter_ids]
-      
-      for f in npyFile:
-        #print(f)
-        inp = np.load(f)
-        if self.transpose:
-          inp = inp.transpose()
-        yield ArrayInput(inp)
-
-  def count_sents(self, filename):
-    '''npzFile = np.load(filename, mmap_mode="r") # for counting sentences, only read the index
-    l = len(npzFile.files)
-    npzFile.close()'''
-    l = 0
-    with open(filename) as f:
-      while True:
-        line = f.readline()
-        if not line:
-          break
-        else:
-          l += 1
-    return l
-     
 class IDReader(BaseTextReader, Serializable):
   """
   Handles the case where we need to read in a single ID (like retrieval problems)
@@ -237,7 +167,6 @@ class IDReader(BaseTextReader, Serializable):
     pass
 
   def read_sents(self, filename, filter_ids=None):
-    #print(filename)
     return map(lambda l: int(l.strip()), self.iterate_filtered(filename, filter_ids))
 
 ###### CorpusParser
