+++ xnmt/specialized_encoders.py
@@ -1,0 +1,110 @@
+import dynet as dy
+
+# This is a file for specialized encoders that implement a particular model
+# Ideally, these will eventually be refactored to use standard components and the ModularEncoder framework,
+#  (for more flexibility), but for ease of implementation it is no problem to perform an initial implementation here.
+
+# This is a CNN-based encoder that was used in the following paper:
+#  http://papers.nips.cc/paper/6186-unsupervised-learning-of-spoken-language-with-visual-context.pdf
+class HarwathSpeechEncoder(Encoder, Serializable):
+  yaml_tag = u'!HarwathSpeechEncoder'
+  def __init__(self, filter_height, filter_width, channels, num_filters, stride):
+    """
+    :param num_layers: depth of the RNN
+    :param input_dim: size of the inputs
+    :param hidden_dim: size of the outputs (and intermediate RNN layer representations)
+    :param model
+    :param rnn_builder_factory: RNNBuilder subclass, e.g. LSTMBuilder
+    """
+    model = model_globals.dynet_param_collection.param_col
+    self.filter_height = filter_height
+    self.filter_width = filter_width
+    self.channels = channels
+    self.num_filters = num_filters
+    self.stride = stride # (2,2)
+
+    normalInit=dy.NormalInitializer(0, 0.1)
+    self.filters1 = model.add_parameters(dim=(self.filter_height[0], self.filter_width[0], self.channels[0], self.num_filters[0]),
+                                         init=normalInit)
+    self.filters2 = model.add_parameters(dim=(self.filter_height[1], self.filter_width[1], self.channels[1], self.num_filters[1]),
+                                         init=normalInit)
+    self.filters3 = model.add_parameters(dim=(self.filter_height[2], self.filter_width[2], self.channels[2], self.num_filters[2]),
+                                         init=normalInit)
+
+  def transduce(self, src):
+    src = src.as_tensor()
+
+    src_height = src.dim()[0][0]
+    src_width = src.dim()[0][1]
+    src_channels = 1
+    batch_size = src.dim()[1]
+
+
+    src = dy.reshape(src, (src_height, src_width, src_channels), batch_size=batch_size) # ((276, 80, 3), 1)
+    # print(self.filters1)
+    # convolution and pooling layers
+    l1 = dy.rectify(dy.conv2d(src, dy.parameter(self.filters1), stride = [self.stride[0], self.stride[0]], is_valid = True))
+    pool1 = dy.maxpooling2d(l1, (1, 4), (1,2), is_valid = True)
+
+    l2 = dy.rectify(dy.conv2d(pool1, dy.parameter(self.filters2), stride = [self.stride[1], self.stride[1]], is_valid = True))
+    pool2 = dy.maxpooling2d(l2, (1, 4), (1,2), is_valid = True)
+
+    l3 = dy.rectify(dy.conv2d(pool2, dy.parameter(self.filters3), stride = [self.stride[2], self.stride[2]], is_valid = True))
+
+    pool3 = dy.max_dim(l3, d = 1)
+    # print(pool3.dim())
+    output = dy.cdiv(pool3,dy.sqrt(dy.squared_norm(pool3)))
+    output = dy.reshape(output, (self.num_filters[2],), batch_size = batch_size)
+    # print("my dim: ", output.dim())
+
+    return ExpressionSequence(expr_tensor=output)
+
+  def initial_state(self):
+    return PseudoState(self)
+
+
+# This is an image encoder that takes in features and does a linear transform from the following paper
+#  http://papers.nips.cc/paper/6186-unsupervised-learning-of-spoken-language-with-visual-context.pdf
+class HarwathImageEncoder(Encoder, Serializable):
+  yaml_tag = u'!HarwathImageEncoder'
+  """
+    Inputs are first put through 2 CNN layers, each with stride (2,2), so dimensionality
+    is reduced by 4 in both directions.
+    Then, we add a configurable number of bidirectional RNN layers on top.
+    """
+
+  def __init__(self, in_height, out_height):
+    """
+      :param num_layers: depth of the RNN
+      :param input_dim: size of the inputs
+      :param hidden_dim: size of the outputs (and intermediate RNN layer representations)
+      :param model
+      :param rnn_builder_factory: RNNBuilder subclass, e.g. LSTMBuilder
+      """
+
+    model = model_globals.dynet_param_collection.param_col
+    self.in_height = in_height
+    self.out_height = out_height
+
+    normalInit=dy.NormalInitializer(0, 0.1)
+    self.pW = model.add_parameters(dim = (self.out_height, self.in_height), init=normalInit)
+    self.pb = model.add_parameters(dim = self.out_height)
+
+  def transduce(self, src):
+    src = src.as_tensor()
+
+    src_height = src.dim()[0][0]
+    src_width = 1
+    batch_size = src.dim()[1]
+
+    W = dy.parameter(self.pW)
+    b = dy.parameter(self.pb)
+
+    src = dy.reshape(src, (src_height, src_width), batch_size=batch_size) # ((276, 80, 3), 1)
+    # convolution and pooling layers
+    l1 = (W*src)+b
+    output = dy.cdiv(l1,dy.sqrt(dy.squared_norm(l1)))
+    return ExpressionSequence(expr_tensor=output)
+
+  def initial_state(self):
+    return PseudoState(self)
